fix filter query

+++ appointment-service/src/main/java/com/docdock/group09/appointment_service/dto/response/AppointmentFilterStatsResponse.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AppointmentFilterStatsResponse {
-    private int numberOfPending;
-    private int numberOfCompleted;
-    private int numberOfCancelled;
-    private int numberOfConfirmed;
+    private long numberOfPending;
+    private long numberOfCompleted;
+    private long numberOfCancelled;
+    private long numberOfConfirmed;
 }
